+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.1] - 2025-01-18
+
+### Added
+- Implemented auto scroll to the current playing song in PlayQueue
+
+### Changed
+- Refresh PlayQueue when clicking on the link even without a page transition
+
 ## [3.1.0] - 2025-01-18
 
 ### Changed